Split dbops imageOps, dataOps, tagOps, and userOps
